SetupFallbacktuner: Support write URL manual
If the remote receiver has HTTP authentication enabled in Enigma,
putting the data in the OpenWebIF parameter is not enough to get past this.
So, enable the ability to write the URL in the correct format
to get past this authentication and be able to play from the remote receiver.