startwizard.xml: Scanning - change next step "reboot" use "end"
Before:
If we performed a search, whether manual or automaticupon completion
we would always return to the first step of the wizard.

Now:
If we don't perform a search, whether manual or automatic
the wizard will launch in OpenspaPlug.
If we perform a manual search, it will ask if we want to perform another one.
If we perform an automatic search
the next step of the wizard will launch in OpenspaPlug.